\name{kevinRpkg-package}
\alias{kevinRpkg-package}
\alias{kevinRpkg}
\docType{package}
\title{\packageTitle{kevinRpkg}}
\description{\packageDescription{kevinRpkg}}
\section{Package Content}{\packageIndices{kevinRpkg}}
\author{\packageAuthor{kevinRpkg}}
\section{Maintainer}{\packageMaintainer{kevinRpkg}}
\keyword{package}
